Migrate from Kobalt to Gradle
