--- kobalt/wrapper/kobalt-wrapper.properties
@@ -1,1 +1,0 @@
-kobalt.version=1.0.114
