+++ settings.gradle.kts
@@ -1,0 +1,3 @@
+rootProject.name = "wave"
+
+include("wave-core", "wave-app")
