+++ wave-app/build.gradle.kts
@@ -1,0 +1,13 @@
+plugins {
+    application
+    kotlin("jvm")
+}
+
+application {
+    mainClassName = "com.github.dhsavell.wave.Launcher"
+}
+
+dependencies {
+    compile(project(":wave-core"))
+    compile(kotlin("stdlib"))
+}
