+++ wave-app/src/main/kotlin/com/github/dhsavell/wave/app/provider/BotCommand.kt
@@ -1,14 +1,12 @@
 package com.github.dhsavell.wave.app.provider
 
 import com.github.dhsavell.wave.core.command.Command
-import net.java.sezpoz.Index
-import net.java.sezpoz.Indexable
 
 @Retention(AnnotationRetention.SOURCE)
 @Target(AnnotationTarget.CLASS, AnnotationTarget.FUNCTION, AnnotationTarget.FIELD)
 @Indexable(type = Command::class)
-annotation class BotListener
+annotation class BotCommand
 
-fun provideListeners(): List<Command> {
+fun provideCommands(): List<Command> {
     return Index.load(BotCommand::class.java, Command::class.java).map { command -> command.instance() }
 }
