+++ wave-core/src/main/kotlin/com/github/dhsavell/wave/core/bot/Bot.kt
