+++ wave-core/src/main/kotlin/com/github/dhsavell/wave/core/command/Category.kt
