+++ wave-core/src/main/kotlin/com/github/dhsavell/wave/core/conversation/ConversationManager.kt
