+++ wave-core/src/main/kotlin/com/github/dhsavell/wave/core/conversation/Prompt.kt
