+++ wave-core/src/main/kotlin/com/github/dhsavell/wave/core/conversation/builder/ConversationBuilder.kt
