+++ wave-core/src/main/kotlin/com/github/dhsavell/wave/core/data/DomainProperty.kt
