+++ wave-core/src/main/kotlin/com/github/dhsavell/wave/core/permission/Permission.kt
