+++ wave-core/src/main/kotlin/com/github/dhsavell/wave/core/util/DslEmbedBuilder.kt
