+++ wave-core/src/test/kotlin/com/github/dhsavell/wave/core/data/IDLinkedPropertyTest.kt
