+++ wave-core/src/test/kotlin/com/github/dhsavell/wave/core/testutil/DummyIDLinkedObject.kt
